AlertDialog showDeleteDialog removed from DeleteCodeDialogFragment. AlertDialog showDeleteDialog moved to DeleteCodeDialogFragment
