+++ app/src/main/java/com/t_ovchinnikova/android/scandroid_2/presentation/DeleteCodeListener.kt
@@ -1,0 +1,5 @@
+package com.t_ovchinnikova.android.scandroid_2.presentation
+
+interface DeleteCodeListener {
+    fun onNoteConfirmed()
+}
