+++ app/src/main/java/com/t_ovchinnikova/android/scandroid_2/presentation/EditCodeNoteListener.kt
@@ -1,4 +1,4 @@
-package com.t_ovchinnikova.android.scandroid_2.presentation.view
+package com.t_ovchinnikova.android.scandroid_2.presentation
 
 interface EditCodeNoteListener {
     fun onNoteConfirmed(note: String)
