+++ app/src/main/res/layout/fragment_scanning.xml
@@ -17,7 +17,7 @@
 
     </androidx.camera.view.PreviewView>
 
-    <com.t_ovchinnikova.android.scandroid_2.presentation.view.ViewFinderOverlay
+    <com.t_ovchinnikova.android.scandroid_2.presentation.views.ViewFinderOverlay
         android:id="@+id/overlay"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
